requirements.txt
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
bleach==3.3.0
cffi==1.14.5
chardet==3.0.4
click==7.1.2
coverage==5.5
coveralls==3.0.1
cox==0.1.post3
dataclasses==0.6
datasets==1.1.3
defusedxml==0.7.1
dill==0.3.3
docopt==0.6.2
entrypoints==0.3
filelock==3.0.12
future==0.18.2
gitdb==4.0.5
gitpython==3.1.11
glm-saga==0.1.1
grpcio==1.34.0
idna==2.10
imageio==2.9.0
importlib-metadata==3.7.3
iniconfig==1.1.1
ipywidgets==7.6.3
jinja2==2.11.3
joblib==0.17.0
jsonschema==3.2.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kornia==0.4.1
lime==0.2.0.1
markupsafe==1.1.1
mistune==0.8.4
multiprocess==0.70.11.1
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
networkx==2.5
notebook==6.2.0
packaging==20.7
pandocfilters==1.4.3
pluggy==0.13.1
prometheus-client==0.9.0
protobuf==3.14.0
psutil==5.7.3
py==1.10.0
py3nvml==0.2.6
pyarrow==2.0.0
pycparser==2.20
pyrsistent==0.17.3
pytest==6.2.4
pytest-mock==3.6.1
pywavelets==1.1.1
regex==2020.11.13
requests==2.25.0
robustness
sacremoses==0.0.43
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.4
send2trash==1.5.0
sentencepiece==0.1.91
smmap==3.0.4
terminado==0.9.3
testpath==0.4.4
tifffile==2020.11.26
tokenizers==0.9.3
toml==0.10.2
torch-lucent==0.1.8
tqdm==4.49.0
transformers==3.5.1
urllib3==1.26.2
webencodings==0.5.1
widgetsnbextension==3.5.1
xmltodict==0.12.0
xxhash==2.0.0
zipp==3.4.1
torchvision==0.8.1
nlp
wordcloud